pdf to image integration

+++ main.py
@@ -7,6 +7,8 @@ from fastapi import FastAPI, Body, Request, Form
 from fastapi import FastAPI, File, UploadFile
 from newpaddle import OCR, parsing
 from paddleocr import PaddleOCR
+from newpaddle import pdfconverter
+from newpaddle import empty_directory
 
 app = FastAPI()
 ocr = PaddleOCR(use_angle_cls=True, lang='en')
@@ -19,29 +21,40 @@ async def root():
 
 @app.post("/scan")
 async def scan_image(request:Request, file:UploadFile):
-    if not os.path.isdir("./image_upload"):
-        os.mkdir("./image_upload")
+    folder = "image-upload"
+    if not os.path.isdir(f"./{folder}"):
+        os.mkdir(f"./{folder}")
         
     file_name=str(file.filename)
-    with open(f"./image_upload/{file_name}", "wb+") as file_object:
+    with open(f"./{folder}/{file_name}", "wb+") as file_object:
         file_object.write(file.file.read())
-    file_path=f"./image_upload/{file_name}"
-    list_of_words = OCR(file_path, ocr)
-    final_results = parsing(list_of_words)
+    file_path=f"./{folder}/{file_name}"
+    if file_name.split(".")[1] == "pdf":
+        pdfconverter(file_path, folder)
+        os.remove(file_path)
+        for image in os.listdir(f"./{folder}"):
+            file_path = f'{folder}/{image}'
 
-    html_content = f"""
-                    <html>
-                        <head>
-                            <title>Results</title>
-                        </head>
-                        <body>
-                            <h1>"{"Scanning Image"}"</h1>
-                        </body>
-                    </html>
-                    """
-       
+    try:
+        list_of_words = OCR(file_path, ocr)
+        final_results = parsing(list_of_words)
+
+        html_content = f"""
+                        <html>
+                            <head>
+                                <title>Results</title>
+                            </head>
+                            <body>
+                                <h1>"{"Scanning Image"}"</h1>
+                            </body>
+                        </html>
+                        """
+        
         
             # return FileResponse('./htmldirectory/df_representation.html')
-    os.remove(file_path)
-    return final_results
+        empty_directory("image_upload")
+        return final_results
+    except:
+        empty_directory("image_upload")
+        return "INVALID IMAGE! TRY AGAIN PLEASE."
